+++ pkg/querier/tenantfederation/merge_queryable.go
@@ -50,7 +50,7 @@ func NewQueryable(upstream storage.Queryable, bypassWithSingleID bool, maxConcur
 			}, nil
 		},
 	}
-	return NewMergeQueryable(defaultTenantLabel, callbacks, bypassWithSingleID, maxConcurrency, reg, logger)
+	return NewMergeQueryable(defaultTenantLabel, callbacks, tenant.NewMultiResolver(), bypassWithSingleID, maxConcurrency, reg, logger)
 }
 
 // MergeQueryableCallbacks contains callbacks to NewMergeQueryable, for customizing its behaviour.
@@ -89,8 +89,9 @@ func (q *tenantQuerier) Close() error {
 	return q.upstream.Close()
 }
 
-// NewMergeQueryable returns a queryable that merges results for all involved federation IDs.
-// The underlying querier is returned by a callback in MergeQueryableCallbacks.
+// NewMergeQueryable returns a queryable that merges results for all involved
+// federation IDs. The underlying querier is returned by a callback in
+// MergeQueryableCallbacks.
 //
 // By setting bypassWithSingleID to true the mergeQuerier gets bypassed,
 // and results for requests with a single ID will not contain the ID label.
@@ -101,11 +102,15 @@ func (q *tenantQuerier) Close() error {
 // If the label `idLabelName` already exists, its value is overwritten and
 // the previous value is exposed through a new label prefixed with "original_".
 // This behaviour is not implemented recursively.
-func NewMergeQueryable(idLabelName string, callbacks MergeQueryableCallbacks, bypassWithSingleID bool, maxConcurrency int, reg prometheus.Registerer, logger log.Logger) storage.Queryable {
+func NewMergeQueryable(idLabelName string, callbacks MergeQueryableCallbacks, resolver tenant.Resolver, bypassWithSingleID bool, maxConcurrency int, reg prometheus.Registerer, logger log.Logger) storage.Queryable {
+	// Note that we allow tenant.Resolver to be injected instead of using the
+	// tenant.TenantIDs() method because GEM needs to inject different behavior
+	// here for the cluster federation feature.
 	return &mergeQueryable{
 		logger:             logger,
 		idLabelName:        idLabelName,
 		callbacks:          callbacks,
+		resolver:           resolver,
 		bypassWithSingleID: bypassWithSingleID,
 		maxConcurrency:     maxConcurrency,
 		tenantsQueried: promauto.With(reg).NewHistogram(prometheus.HistogramOpts{
@@ -121,6 +126,7 @@ type mergeQueryable struct {
 	idLabelName        string
 	bypassWithSingleID bool
 	callbacks          MergeQueryableCallbacks
+	resolver           tenant.Resolver
 	maxConcurrency     int
 	tenantsQueried     prometheus.Histogram
 }
@@ -136,6 +142,7 @@ func (m *mergeQueryable) Querier(mint int64, maxt int64) (storage.Querier, error
 		logger:             m.logger,
 		idLabelName:        m.idLabelName,
 		callbacks:          m.callbacks,
+		resolver:           m.resolver,
 		upstream:           upstream,
 		maxConcurrency:     m.maxConcurrency,
 		bypassWithSingleID: m.bypassWithSingleID,
@@ -151,6 +158,7 @@ func (m *mergeQueryable) Querier(mint int64, maxt int64) (storage.Querier, error
 type mergeQuerier struct {
 	logger             log.Logger
 	callbacks          MergeQueryableCallbacks
+	resolver           tenant.Resolver
 	upstream           MergeQuerierUpstream
 	idLabelName        string
 	maxConcurrency     int
@@ -164,7 +172,7 @@ type mergeQuerier struct {
 // For the label "original_" + `idLabelName it will return all values
 // for the original `idLabelName` label.
 func (m *mergeQuerier) LabelValues(ctx context.Context, name string, matchers ...*labels.Matcher) ([]string, annotations.Annotations, error) {
-	ids, err := tenant.TenantIDs(ctx)
+	ids, err := m.resolver.TenantIDs(ctx)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -203,7 +211,7 @@ func (m *mergeQuerier) LabelValues(ctx context.Context, name string, matchers ..
 // LabelNames returns all the unique label names present for involved federation IDs.
 // It also adds the `idLabelName` and if present in the original results the original `idLabelName`.
 func (m *mergeQuerier) LabelNames(ctx context.Context, matchers ...*labels.Matcher) ([]string, annotations.Annotations, error) {
-	ids, err := tenant.TenantIDs(ctx)
+	ids, err := m.resolver.TenantIDs(ctx)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -314,7 +322,7 @@ func (m *mergeQuerier) Close() error {
 // If the `idLabelName` is matched on, it only considers matching IDs.
 // The forwarded labelSelector does not contain those that operate on `idLabelName`.
 func (m *mergeQuerier) Select(ctx context.Context, sortSeries bool, hints *storage.SelectHints, matchers ...*labels.Matcher) storage.SeriesSet {
-	ids, err := tenant.TenantIDs(ctx)
+	ids, err := m.resolver.TenantIDs(ctx)
 	if err != nil {
 		return storage.ErrSeriesSet(err)
 	}