+++ pkg/util/noauth/no_auth.go
@@ -3,23 +3,30 @@
 // Provenance-includes-license: Apache-2.0
 // Provenance-includes-copyright: The Cortex Authors.
 
-// Package noauth provides middlewares thats injects a tenant ID, so the rest of the code
-// can continue to be multitenant.
+// Package noauth provides middlewares that injects a tenant ID so the rest of the code
+// can continue to be multitenant and validates the number of tenant IDs if supplied.
 package noauth
 
 import (
 	"context"
+	"fmt"
 	"net/http"
 
 	"github.com/grafana/dskit/middleware"
 	"github.com/grafana/dskit/server"
+	"github.com/grafana/dskit/tenant"
 	"github.com/grafana/dskit/user"
 	"google.golang.org/grpc"
 )
 
-// SetupAuthMiddleware for the given server config.
-func SetupAuthMiddleware(config *server.Config, enabled bool, noGRPCAuthOn []string, noAuthTenant string) middleware.Interface {
-	if enabled {
+const (
+	tooManyTenantsTemplate = "too many tenant IDs present. max: %d actual: %d"
+	noTenantIDTemplate     = "no tenant ID present. set tenant ID using the %s header"
+)
+
+// SetupTenantMiddleware for the given server config.
+func SetupTenantMiddleware(config *server.Config, multitenancyEnabled bool, noMultitenancyTenant string, federationEnabled bool, federationMaxTenants int, noGRPCAuthOn []string) middleware.Interface {
+	if multitenancyEnabled {
 		ignoredMethods := map[string]bool{}
 		for _, m := range noGRPCAuthOn {
 			ignoredMethods[m] = true
@@ -41,18 +48,18 @@ func SetupAuthMiddleware(config *server.Config, enabled bool, noGRPCAuthOn []str
 			},
 		)
 
-		return middleware.AuthenticateUser
+		return newTenantValidationMiddleware(federationEnabled, federationMaxTenants)
 	}
 
 	config.GRPCMiddleware = append(config.GRPCMiddleware,
 		func(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
-			ctx = user.InjectOrgID(ctx, noAuthTenant)
+			ctx = user.InjectOrgID(ctx, noMultitenancyTenant)
 			return handler(ctx, req)
 		},
 	)
 	config.GRPCStreamMiddleware = append(config.GRPCStreamMiddleware,
 		func(srv interface{}, ss grpc.ServerStream, _ *grpc.StreamServerInfo, handler grpc.StreamHandler) error {
-			ctx := user.InjectOrgID(ss.Context(), noAuthTenant)
+			ctx := user.InjectOrgID(ss.Context(), noMultitenancyTenant)
 
 			return handler(srv, serverStream{
 				ctx:          ctx,
@@ -62,11 +69,42 @@ func SetupAuthMiddleware(config *server.Config, enabled bool, noGRPCAuthOn []str
 	)
 	return middleware.Func(func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			ctx := user.InjectOrgID(r.Context(), noAuthTenant)
+			ctx := user.InjectOrgID(r.Context(), noMultitenancyTenant)
 			next.ServeHTTP(w, r.WithContext(ctx))
 		})
 	})
 
+}
+
+func newTenantValidationMiddleware(federation bool, maxTenants int) middleware.Interface {
+	return middleware.Func(func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			_, ctx, err := user.ExtractOrgIDFromHTTPRequest(r)
+			if err != nil {
+				http.Error(w, fmt.Sprintf(noTenantIDTemplate, user.OrgIDHeaderName), http.StatusUnauthorized)
+				return
+			}
+
+			ids, err := tenant.TenantIDs(ctx)
+			if err != nil {
+				http.Error(w, err.Error(), http.StatusUnauthorized)
+				return
+			}
+
+			numIds := len(ids)
+			if !federation && numIds > 1 {
+				http.Error(w, fmt.Sprintf(tooManyTenantsTemplate, 1, numIds), http.StatusUnprocessableEntity)
+				return
+			}
+
+			if federation && maxTenants > 0 && numIds > maxTenants {
+				http.Error(w, fmt.Sprintf(tooManyTenantsTemplate, maxTenants, numIds), http.StatusUnprocessableEntity)
+				return
+			}
+
+			next.ServeHTTP(w, r.WithContext(ctx))
+		})
+	})
 }
 
 type serverStream struct {